backend/requirements.txt
Django==4.2.3
python-dotenv==1.0.0
djangorestframework==3.14.0
pytest==7.4.0
pytest-django==4.5.2
black==23.7.0
flake8==6.0.0
django-mptt==0.14.0
drf-spectacular==0.26.3
coverage==7.2.7
pytest-cov==4.1.0
Pillow==10.0.0